+++ ProjetoDWM/database/migrations/2020_10_23_143143_create_comentarios_table.php
@@ -14,7 +14,7 @@ class CreateComentariosTable extends Migration
     public function up()
     {
         Schema::create('comentarios', function (Blueprint $table) {
-            $table->bigInteger('id');
+            $table->serial('id');
             $table->bigInteger('id_video');
             $table->String('nome');
             $table->String('comentario');
